Changer for post and REST implementation, async responses WIP
