+++ .env
@@ -1,0 +1,3 @@
+PKEY=1234567890
+FS_IP=localhost
+FS_PORT=8082
