+++ .gitignore
@@ -1,12 +1,15 @@
 #test file
 #test.js
-.env
 # Logs
 logs
 *.log
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+/encrypted
+/input
+/output
 
 # Runtime data
 pids
@@ -58,7 +61,7 @@ typings/
 .yarn-integrity
 
 # dotenv environment variables file
-.env
+#.env
 
 # next.js build output
 .next
