--- input/hola.txt
@@ -1,1 +1,0 @@
-Hola rossi sos un cagon
