+++ package.json
@@ -5,7 +5,9 @@
   "main": "test.js",
   "dependencies": {
     "dotenv": "^8.1.0",
-    "express": "^4.17.1"
+    "express": "^4.17.1",
+    "hbs": "^4.0.6",
+    "multer": "^1.4.2"
   },
   "devDependencies": {},
   "scripts": {
